% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_courses.R
\name{load_course_demo}
\alias{load_course_demo}
\title{demo load course function
This function will load and install a the mbb343 course}
\usage{
load_course_demo()
}
\description{
demo load course function
This function will load and install a the mbb343 course
}
\examples{
load_course_demo()
}
